+++ src/api/instance/defaultInstance.ts
@@ -1,4 +1,3 @@
-import { TOKEN_KEY } from "@/store/useTokenStore";
 import axios from "axios";
 
 const baseURL = `${import.meta.env.VITE_API_URL}`;
@@ -9,18 +8,5 @@ const defaultInstance = axios.create({
     "Content-Type": "application/json",
   },
 });
-
-defaultInstance.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem(TOKEN_KEY);
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-    return config;
-  },
-  (error) => {
-    return Promise.reject(error);
-  },
-);
 
 export default defaultInstance;
