+++ src/api/login.ts
@@ -1,3 +1,4 @@
+import authInstance from "./instance/authInstance";
 import defaultInstance from "./instance/defaultInstance";
 
 export const getKakaoUrl = async () => {
@@ -28,7 +29,7 @@ export const postLogin = async (kakaoId: number) => {
 };
 
 export const postRefresh = async () => {
-  const response = await defaultInstance.post("/auth/refresh", null, {
+  const response = await authInstance.post("/auth/refresh", null, {
     withCredentials: true,
   });
   return response.data;
